Add files via upload
latvian_translation - Scrapes lsm.lv headlines and retrieves summary of the news item. It then breaks down the summary into words and uses Gtranslate api to translate to EN. The translated words are saved to the latvian_translation.csv which is ever growing, on a daily basis (code initially fetches csv, converts to df, and then to dict and checks for duplicate keys and adds only unique keys). Ideal for building up vocabulary